Move pkg/logexec → pkg/dexec
