+++ pkg/dexec/borrowed_exec_test.go
@@ -9,7 +9,7 @@
 // circular dependency on non-cgo darwin.
 
 //nolint
-package logexec_test
+package dexec_test
 
 import (
 	"bufio"
@@ -28,8 +28,8 @@ import (
 	"testing"
 	"time"
 
+	exec "github.com/datawire/teleproxy/pkg/dexec"
 	"github.com/datawire/teleproxy/pkg/dlog"
-	exec "github.com/datawire/teleproxy/pkg/logexec"
 )
 
 func testContext(t *testing.T) context.Context {
