+++ pkg/dexec/borrowed_internal_test.go
@@ -5,7 +5,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-package logexec
+package dexec
 
 import (
 	"io"
