+++ pkg/dexec/reader.go
@@ -1,4 +1,4 @@
-package logexec
+package dexec
 
 import (
 	"bytes"
